+++ src/providers/index.ts
@@ -1,3 +1,3 @@
 export { ForceCodeContentProvider } from './ContentProvider';
-export { ForceCodeLogProvider, filterLog } from './LogProvider';
+export { filterLog } from './LogProvider';
 export { ApexTestLinkProvider } from './ApexTestLinkProvider';